refactor(controller): Improve listeners and servers shutdown
These changes were introduced in the same vein of the changes to
`startListeners`. In the future, we will want to have a specific
shutdown order to the servers/listeners.

Additionally, these changes decouple the Controller and the Worker
shutdown. Currently, a call to `stopListeners` on an instance that is
running as both a Controller and a Worker would stop the listeners
for both.