feat(controller): Enable support for `ops` purposed listeners
This change implements a new listener purpose, `ops`.

`ops`, short for operations, is the new listener that will expose the
health endpoint, and potentially metrics too.

An `ops` purposed listener is responsible for serving Boundary's
operations endpoints, like health, and possibly metrics.

Startup and shutdown: At startup, `ops` listeners get configured as
late as possible to catch any errors on the startup;
At shutdown, `ops` listeners (if present), introduce a grace period
delay in the process. This is to enable the health endpoint to report
unhealthy for that grace period. They are also the last listeners to be
shutdown.